Adding a warning to use multi-workers on FnApiRunner
